+++ examples/opensora_pku/opensora/models/diffusion/latte/modeling_latte.py
@@ -4,8 +4,10 @@ import logging
 import os
 from typing import Any, Dict, Optional
 
+from opensora.models.diffusion.utils.pos_embed import PositionGetter1D, PositionGetter2D, get_1d_sincos_pos_embed
+
 import mindspore as ms
-from mindspore import Parameter, nn, ops
+from mindspore import nn, ops
 
 from mindone.diffusers.configuration_utils import ConfigMixin, register_to_config
 
@@ -13,1096 +15,16 @@ from mindone.diffusers.configuration_utils import ConfigMixin, register_to_confi
 from mindone.diffusers.models.modeling_utils import ModelMixin
 
 from .modules import (
-    AdaLayerNorm,
     AdaLayerNormSingle,
-    AdaLayerNormZero,
+    BasicTransformerBlock,
+    BasicTransformerBlock_,
     CaptionProjection,
-    FeedForward,
-    GatedSelfAttentionDense,
-    ImagePositionalEmbeddings,
+    LatteT2VBlock,
     LayerNorm,
-    MultiHeadAttention,
     PatchEmbed,
-    SinusoidalPositionalEmbedding,
 )
-from .pos import get_1d_sincos_pos_embed
-
-# from mindone.diffusers.models.normalization import AdaLayerNorm, AdaLayerNormZero
-
-
-# from mindone.diffusers.models.lora import LoRACompatibleConv, LoRACompatibleLinear
 
 logger = logging.getLogger(__name__)
-
-
-class BasicTransformerBlock_(nn.Cell):
-    r"""
-    A basic Transformer block.
-
-    Parameters:
-        dim (`int`): The number of channels in the input and output.
-        num_attention_heads (`int`): The number of heads to use for multi-head attention.
-        attention_head_dim (`int`): The number of channels in each head.
-        dropout (`float`, *optional*, defaults to 0.0): The dropout probability to use.
-        cross_attention_dim (`int`, *optional*): The size of the encoder_hidden_states vector for cross attention.
-        activation_fn (`str`, *optional*, defaults to `"geglu"`): Activation function to be used in feed-forward.
-        num_embeds_ada_norm (:
-            obj: `int`, *optional*): The number of diffusion steps used during training. See `Transformer2DModel`.
-        attention_bias (:
-            obj: `bool`, *optional*, defaults to `False`): Configure if the attentions should contain a bias parameter.
-        only_cross_attention (`bool`, *optional*):
-            Whether to use only cross-attention layers. In this case two cross attention layers are used.
-        double_self_attention (`bool`, *optional*):
-            Whether to use two self-attention layers. In this case no cross attention layers are used.
-        upcast_attention (`bool`, *optional*):
-            Whether to upcast the attention computation to float32. This is useful for mixed precision training.
-        norm_elementwise_affine (`bool`, *optional*, defaults to `True`):
-            Whether to use learnable elementwise affine parameters for normalization.
-        norm_type (`str`, *optional*, defaults to `"layer_norm"`):
-            The normalization layer to use. Can be `"layer_norm"`, `"ada_norm"` or `"ada_norm_zero"`.
-        final_dropout (`bool` *optional*, defaults to False):
-            Whether to apply a final dropout after the last feed-forward layer.
-        attention_type (`str`, *optional*, defaults to `"default"`):
-            The type of attention to use. Can be `"default"` or `"gated"` or `"gated-text-image"`.
-        positional_embeddings (`str`, *optional*, defaults to `None`):
-            The type of positional embeddings to apply to.
-        num_positional_embeddings (`int`, *optional*, defaults to `None`):
-            The maximum number of positional embeddings to apply.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        num_attention_heads: int,
-        attention_head_dim: int,
-        dropout=0.0,
-        cross_attention_dim: Optional[int] = None,
-        activation_fn: str = "geglu",
-        num_embeds_ada_norm: Optional[int] = None,
-        attention_bias: bool = False,
-        only_cross_attention: bool = False,
-        double_self_attention: bool = False,
-        upcast_attention: bool = False,
-        norm_elementwise_affine: bool = True,
-        norm_type: str = "layer_norm",  # 'layer_norm', 'ada_norm', 'ada_norm_zero', 'ada_norm_single'
-        norm_eps: float = 1e-5,
-        final_dropout: bool = False,
-        attention_type: str = "default",
-        positional_embeddings: Optional[str] = None,
-        num_positional_embeddings: Optional[int] = None,
-        enable_flash_attention: bool = False,
-    ):
-        super().__init__()
-        self.only_cross_attention = only_cross_attention
-
-        self.use_ada_layer_norm_zero = (num_embeds_ada_norm is not None) and norm_type == "ada_norm_zero"
-        self.use_ada_layer_norm = (num_embeds_ada_norm is not None) and norm_type == "ada_norm"
-        self.use_ada_layer_norm_single = norm_type == "ada_norm_single"
-        self.use_layer_norm = norm_type == "layer_norm"
-
-        if norm_type in ("ada_norm", "ada_norm_zero") and num_embeds_ada_norm is None:
-            raise ValueError(
-                f"`norm_type` is set to {norm_type}, but `num_embeds_ada_norm` is not defined. Please make sure to"
-                f" define `num_embeds_ada_norm` if setting `norm_type` to {norm_type}."
-            )
-
-        if positional_embeddings and (num_positional_embeddings is None):
-            raise ValueError(
-                "If `positional_embedding` type is defined, `num_positition_embeddings` must also be defined."
-            )
-
-        if positional_embeddings == "sinusoidal":
-            self.pos_embed = SinusoidalPositionalEmbedding(dim, max_seq_length=num_positional_embeddings)
-        else:
-            self.pos_embed = None
-
-        # Define 3 blocks. Each block has its own normalization layer.
-        # 1. Self-Attn
-        if self.use_ada_layer_norm:
-            self.norm1_ada = AdaLayerNorm(dim, num_embeds_ada_norm)
-            self.norm1_ada.norm = LayerNorm(dim, elementwise_affine=False)
-        elif self.use_ada_layer_norm_zero:
-            self.norm1_ada_zero = AdaLayerNormZero(dim, num_embeds_ada_norm)
-            self.norm1_ada_zero.norm = LayerNorm(dim, elementwise_affine=False)
-        else:
-            self.norm1_ln = LayerNorm(dim, elementwise_affine=norm_elementwise_affine, eps=norm_eps)
-
-        self.attn1 = MultiHeadAttention(
-            query_dim=dim,
-            heads=num_attention_heads,
-            dim_head=attention_head_dim,
-            dropout=dropout,
-            bias=attention_bias,
-            cross_attention_dim=cross_attention_dim if only_cross_attention else None,
-            upcast_attention=upcast_attention,
-            enable_flash_attention=enable_flash_attention,
-        )
-
-        self.norm3 = LayerNorm(dim, elementwise_affine=norm_elementwise_affine, eps=norm_eps)
-
-        self.ff = FeedForward(dim, dropout=dropout, activation_fn=activation_fn, final_dropout=final_dropout)
-
-        # 4. Fuser
-        if attention_type == "gated" or attention_type == "gated-text-image":
-            self.fuser = GatedSelfAttentionDense(dim, cross_attention_dim, num_attention_heads, attention_head_dim)
-
-        # 5. Scale-shift for PixArt-Alpha.
-        if self.use_ada_layer_norm_single:
-            self.scale_shift_table = ms.Parameter(ops.randn(6, dim) / dim**0.5)
-
-        # let chunk size default to None
-        self._chunk_size = None
-        self._chunk_dim = 0
-
-    def construct(
-        self,
-        hidden_states: ms.Tensor,
-        attention_mask: Optional[ms.Tensor] = None,
-        encoder_hidden_states: Optional[ms.Tensor] = None,
-        encoder_attention_mask: Optional[ms.Tensor] = None,
-        timestep: Optional[ms.Tensor] = None,
-        cross_attention_kwargs: Dict[str, Any] = None,
-        class_labels: Optional[ms.Tensor] = None,
-    ) -> ms.Tensor:
-        # Notice that normalization is always applied before the real computation in the following blocks.
-        # 0. Self-Attention
-        batch_size = hidden_states.shape[0]
-
-        gate_msa, shift_mlp, scale_mlp, gate_mlp = None, None, None, None
-        if self.use_ada_layer_norm:
-            norm_hidden_states = self.norm1_ada(hidden_states, timestep)
-        elif self.use_ada_layer_norm_zero:
-            norm_hidden_states, gate_msa, shift_mlp, scale_mlp, gate_mlp = self.norm1_ada_zero(
-                hidden_states, timestep, class_labels, hidden_dtype=hidden_states.dtype
-            )
-        elif self.use_layer_norm:
-            norm_hidden_states = self.norm1_ln(hidden_states)
-        elif self.use_ada_layer_norm_single:
-            shift_msa, scale_msa, gate_msa, shift_mlp, scale_mlp, gate_mlp = (
-                self.scale_shift_table[None] + timestep.reshape(batch_size, 6, -1)
-            ).chunk(6, axis=1)
-            norm_hidden_states = self.norm1_ln(hidden_states)
-            norm_hidden_states = norm_hidden_states * (1 + scale_msa) + shift_msa
-            # norm_hidden_states = norm_hidden_states.squeeze(1)  # error message
-        else:
-            raise ValueError("Incorrect norm used")
-
-        if self.pos_embed is not None:
-            norm_hidden_states = self.pos_embed(norm_hidden_states)
-
-        # 1. Retrieve lora scale.
-        lora_scale = cross_attention_kwargs.get("scale", 1.0) if cross_attention_kwargs is not None else 1.0
-
-        # 2. Prepare GLIGEN inputs
-        cross_attention_kwargs = cross_attention_kwargs.copy() if cross_attention_kwargs is not None else {}
-        if "gligen" in cross_attention_kwargs:
-            gligen_kwargs = cross_attention_kwargs["gligen"]
-            # del cross_attention_kwargs["gligen"]
-        else:
-            gligen_kwargs = None
-        attn_output = self.attn1(
-            norm_hidden_states,
-            encoder_hidden_states=encoder_hidden_states if self.only_cross_attention else None,
-            attention_mask=attention_mask,
-            **cross_attention_kwargs,
-        )
-        if self.use_ada_layer_norm_zero:
-            attn_output = gate_msa.unsqueeze(1) * attn_output
-        elif self.use_ada_layer_norm_single:
-            attn_output = gate_msa * attn_output
-
-        hidden_states = attn_output + hidden_states
-        if hidden_states.ndim == 4:
-            hidden_states = hidden_states.squeeze(1)
-
-        # 2.5 GLIGEN Control
-        if gligen_kwargs is not None:
-            hidden_states = self.fuser(hidden_states, gligen_kwargs["objs"])
-
-        if self.use_ada_layer_norm_zero:
-            norm_hidden_states = norm_hidden_states * (1 + scale_mlp[:, None]) + shift_mlp[:, None]
-
-        if self.use_ada_layer_norm_single:
-            # norm_hidden_states = self.norm2(hidden_states)
-            norm_hidden_states = self.norm3(hidden_states)
-            norm_hidden_states = norm_hidden_states * (1 + scale_mlp) + shift_mlp
-
-        if self._chunk_size is not None:
-            # "feed_forward_chunk_size" can be used to save memory
-            if norm_hidden_states.shape[self._chunk_dim] % self._chunk_size != 0:
-                raise ValueError(
-                    f"`hidden_states` dimension to be chunked: {norm_hidden_states.shape[self._chunk_dim]}"
-                    f"has to be divisible by chunk size: {self._chunk_size}. Make sure to set an"
-                    f"appropriate `chunk_size` when calling `unet.enable_forward_chunking`."
-                )
-
-            num_chunks = norm_hidden_states.shape[self._chunk_dim] // self._chunk_size
-            ff_output = ops.cat(
-                [
-                    self.ff(hid_slice, scale=lora_scale)
-                    for hid_slice in norm_hidden_states.chunk(num_chunks, axis=self._chunk_dim)
-                ],
-                dim=self._chunk_dim,
-            )
-        else:
-            ff_output = self.ff(norm_hidden_states, scale=lora_scale)
-
-        if self.use_ada_layer_norm_zero:
-            ff_output = gate_mlp.unsqueeze(1) * ff_output
-        elif self.use_ada_layer_norm_single:
-            ff_output = gate_mlp * ff_output
-
-        hidden_states = ff_output + hidden_states
-        if hidden_states.ndim == 4:
-            hidden_states = hidden_states.squeeze(1)
-
-        return hidden_states
-
-
-class BasicTransformerBlock(nn.Cell):
-    r"""
-    A basic Transformer block.
-
-    Parameters:
-        dim (`int`): The number of channels in the input and output.
-        num_attention_heads (`int`): The number of heads to use for multi-head attention.
-        attention_head_dim (`int`): The number of channels in each head.
-        dropout (`float`, *optional*, defaults to 0.0): The dropout probability to use.
-        cross_attention_dim (`int`, *optional*): The size of the encoder_hidden_states vector for cross attention.
-        activation_fn (`str`, *optional*, defaults to `"geglu"`): Activation function to be used in feed-forward.
-        num_embeds_ada_norm (:
-            obj: `int`, *optional*): The number of diffusion steps used during training. See `Transformer2DModel`.
-        attention_bias (:
-            obj: `bool`, *optional*, defaults to `False`): Configure if the attentions should contain a bias parameter.
-        only_cross_attention (`bool`, *optional*):
-            Whether to use only cross-attention layers. In this case two cross attention layers are used.
-        double_self_attention (`bool`, *optional*):
-            Whether to use two self-attention layers. In this case no cross attention layers are used.
-        upcast_attention (`bool`, *optional*):
-            Whether to upcast the attention computation to float32. This is useful for mixed precision training.
-        norm_elementwise_affine (`bool`, *optional*, defaults to `True`):
-            Whether to use learnable elementwise affine parameters for normalization.
-        norm_type (`str`, *optional*, defaults to `"layer_norm"`):
-            The normalization layer to use. Can be `"layer_norm"`, `"ada_norm"` or `"ada_norm_zero"`.
-        final_dropout (`bool` *optional*, defaults to False):
-            Whether to apply a final dropout after the last feed-forward layer.
-        attention_type (`str`, *optional*, defaults to `"default"`):
-            The type of attention to use. Can be `"default"` or `"gated"` or `"gated-text-image"`.
-        positional_embeddings (`str`, *optional*, defaults to `None`):
-            The type of positional embeddings to apply to.
-        num_positional_embeddings (`int`, *optional*, defaults to `None`):
-            The maximum number of positional embeddings to apply.
-    """
-
-    def __init__(
-        self,
-        dim: int,
-        num_attention_heads: int,
-        attention_head_dim: int,
-        dropout=0.0,
-        cross_attention_dim: Optional[int] = None,
-        activation_fn: str = "geglu",
-        num_embeds_ada_norm: Optional[int] = None,
-        attention_bias: bool = False,
-        only_cross_attention: bool = False,
-        double_self_attention: bool = False,
-        upcast_attention: bool = False,
-        norm_elementwise_affine: bool = True,
-        norm_type: str = "layer_norm",  # 'layer_norm', 'ada_norm', 'ada_norm_zero', 'ada_norm_single'
-        norm_eps: float = 1e-5,
-        final_dropout: bool = False,
-        attention_type: str = "default",
-        positional_embeddings: Optional[str] = None,
-        num_positional_embeddings: Optional[int] = None,
-        enable_flash_attention: bool = False,
-    ):
-        super().__init__()
-        self.only_cross_attention = only_cross_attention
-
-        self.use_ada_layer_norm_zero = (num_embeds_ada_norm is not None) and norm_type == "ada_norm_zero"
-        self.use_ada_layer_norm = (num_embeds_ada_norm is not None) and norm_type == "ada_norm"
-        self.use_ada_layer_norm_single = norm_type == "ada_norm_single"
-        self.use_layer_norm = norm_type == "layer_norm"
-
-        if norm_type in ("ada_norm", "ada_norm_zero") and num_embeds_ada_norm is None:
-            raise ValueError(
-                f"`norm_type` is set to {norm_type}, but `num_embeds_ada_norm` is not defined. Please make sure to"
-                f" define `num_embeds_ada_norm` if setting `norm_type` to {norm_type}."
-            )
-
-        if positional_embeddings and (num_positional_embeddings is None):
-            raise ValueError(
-                "If `positional_embedding` type is defined, `num_positition_embeddings` must also be defined."
-            )
-
-        if positional_embeddings == "sinusoidal":
-            self.pos_embed = SinusoidalPositionalEmbedding(dim, max_seq_length=num_positional_embeddings)
-        else:
-            self.pos_embed = None
-
-        # Define 3 blocks. Each block has its own normalization layer.
-        # 1. Self-Attn
-        if self.use_ada_layer_norm:
-            self.norm1_ada = AdaLayerNorm(dim, num_embeds_ada_norm)
-            self.norm1_ada.norm = LayerNorm(dim, elementwise_affine=False)
-        elif self.use_ada_layer_norm_zero:
-            self.norm1_ada_zero = AdaLayerNormZero(dim, num_embeds_ada_norm)
-            self.norm1_ada_zero.norm = LayerNorm(dim, elementwise_affine=False)
-        else:
-            self.norm1_ln = LayerNorm(dim, elementwise_affine=norm_elementwise_affine, eps=norm_eps)
-
-        self.attn1 = MultiHeadAttention(
-            query_dim=dim,
-            heads=num_attention_heads,
-            dim_head=attention_head_dim,
-            dropout=dropout,
-            bias=attention_bias,
-            cross_attention_dim=cross_attention_dim if only_cross_attention else None,
-            upcast_attention=upcast_attention,
-            enable_flash_attention=enable_flash_attention,
-        )
-
-        # 2. Cross-Attn
-        if cross_attention_dim is not None or double_self_attention:
-            # We currently only use AdaLayerNormZero for self attention where there will only be one attention block.
-            # I.e. the number of returned modulation chunks from AdaLayerZero would not make sense if returned during
-            # the second cross attention block.
-            if self.use_ada_layer_norm:
-                self.norm2_ada = AdaLayerNorm(dim, num_embeds_ada_norm)
-                self.norm2_ada.norm = LayerNorm(dim, elementwise_affine=False)
-            else:
-                self.norm2_ln = LayerNorm(dim, elementwise_affine=norm_elementwise_affine, eps=norm_eps)
-
-            self.attn2 = MultiHeadAttention(
-                query_dim=dim,
-                cross_attention_dim=cross_attention_dim if not double_self_attention else None,
-                heads=num_attention_heads,
-                dim_head=attention_head_dim,
-                dropout=dropout,
-                bias=attention_bias,
-                upcast_attention=upcast_attention,
-                enable_flash_attention=enable_flash_attention,
-            )  # is self-attn if encoder_hidden_states is none
-        else:
-            self.norm2 = None
-            self.attn2 = None
-
-        # 3. Feed-forward
-        if not self.use_ada_layer_norm_single:
-            self.norm3 = LayerNorm(dim, elementwise_affine=norm_elementwise_affine, eps=norm_eps)
-
-        self.ff = FeedForward(
-            dim,
-            dropout=dropout,
-            activation_fn=activation_fn,
-            final_dropout=final_dropout,
-        )
-
-        # 4. Fuser
-        if attention_type == "gated" or attention_type == "gated-text-image":
-            self.fuser = GatedSelfAttentionDense(dim, cross_attention_dim, num_attention_heads, attention_head_dim)
-
-        # 5. Scale-shift for PixArt-Alpha.
-        if self.use_ada_layer_norm_single:
-            self.scale_shift_table = ms.Parameter(ops.randn(6, dim) / dim**0.5)
-
-        # let chunk size default to None
-        self._chunk_size = None
-        self._chunk_dim = 0
-
-    def construct(
-        self,
-        hidden_states: ms.Tensor,
-        attention_mask: Optional[ms.Tensor] = None,
-        encoder_hidden_states: Optional[ms.Tensor] = None,
-        encoder_attention_mask: Optional[ms.Tensor] = None,
-        timestep: Optional[ms.Tensor] = None,
-        cross_attention_kwargs: Dict[str, Any] = None,
-        class_labels: Optional[ms.Tensor] = None,
-    ) -> ms.Tensor:
-        # Notice that normalization is always applied before the real computation in the following blocks.
-        # 0. Self-Attention
-        batch_size = hidden_states.shape[0]
-        gate_msa, shift_mlp, scale_mlp, gate_mlp = None, None, None, None
-        if self.use_ada_layer_norm:
-            norm_hidden_states = self.norm1_ada(hidden_states, timestep)
-        elif self.use_ada_layer_norm_zero:
-            norm_hidden_states, gate_msa, shift_mlp, scale_mlp, gate_mlp = self.norm1_ada_zero(
-                hidden_states, timestep, class_labels, hidden_dtype=hidden_states.dtype
-            )
-        elif self.use_layer_norm:
-            norm_hidden_states = self.norm1_ln(hidden_states)
-        elif self.use_ada_layer_norm_single:
-            shift_msa, scale_msa, gate_msa, shift_mlp, scale_mlp, gate_mlp = (
-                self.scale_shift_table[None] + timestep.reshape(batch_size, 6, -1)
-            ).chunk(6, axis=1)
-            norm_hidden_states = self.norm1_ln(hidden_states)
-            norm_hidden_states = norm_hidden_states * (1 + scale_msa) + shift_msa
-            # norm_hidden_states = norm_hidden_states.squeeze(1)  # error message
-        else:
-            raise ValueError("Incorrect norm used")
-
-        if self.pos_embed is not None:
-            norm_hidden_states = self.pos_embed(norm_hidden_states)
-
-        # 1. Retrieve lora scale.
-        lora_scale = cross_attention_kwargs.get("scale", 1.0) if cross_attention_kwargs is not None else 1.0
-
-        # 2. Prepare GLIGEN inputs
-        cross_attention_kwargs = cross_attention_kwargs.copy() if cross_attention_kwargs is not None else {}
-        if "gligen" in cross_attention_kwargs:
-            gligen_kwargs = cross_attention_kwargs["gligen"]
-            # del cross_attention_kwargs["gligen"]
-        else:
-            gligen_kwargs = None
-        attn_output = self.attn1(
-            norm_hidden_states,
-            encoder_hidden_states=encoder_hidden_states if self.only_cross_attention else None,
-            attention_mask=attention_mask,
-            **cross_attention_kwargs,
-        )
-        if self.use_ada_layer_norm_zero:
-            attn_output = gate_msa.unsqueeze(1) * attn_output
-        elif self.use_ada_layer_norm_single:
-            attn_output = gate_msa * attn_output
-
-        hidden_states = attn_output + hidden_states
-        if hidden_states.ndim == 4:
-            hidden_states = hidden_states.squeeze(1)
-
-        # 2.5 GLIGEN Control
-        if gligen_kwargs is not None:
-            hidden_states = self.fuser(hidden_states, gligen_kwargs["objs"])
-
-        # 3. Cross-Attention
-        if self.attn2 is not None:
-            if self.use_ada_layer_norm:
-                norm_hidden_states = self.norm2_ada(hidden_states, timestep)
-            elif self.use_ada_layer_norm_zero or self.use_layer_norm:
-                norm_hidden_states = self.norm2_ln(hidden_states)
-            elif self.use_ada_layer_norm_single:
-                # For PixArt norm2 isn't applied here:
-                # https://github.com/PixArt-alpha/PixArt-alpha/blob/0f55e922376d8b797edd44d25d0e7464b260dcab/diffusion/model/nets/PixArtMS.py#L70C1-L76C103
-                norm_hidden_states = hidden_states
-            else:
-                raise ValueError("Incorrect norm")
-
-            if self.pos_embed is not None and self.use_ada_layer_norm_single is False:
-                norm_hidden_states = self.pos_embed(norm_hidden_states)
-
-            attn_output = self.attn2(
-                norm_hidden_states,
-                encoder_hidden_states=encoder_hidden_states,
-                attention_mask=encoder_attention_mask,
-                **cross_attention_kwargs,
-            )
-            hidden_states = attn_output + hidden_states
-
-        # 4. Feed-forward
-        if not self.use_ada_layer_norm_single:
-            norm_hidden_states = self.norm3(hidden_states)
-
-        if self.use_ada_layer_norm_zero:
-            norm_hidden_states = norm_hidden_states * (1 + scale_mlp[:, None]) + shift_mlp[:, None]
-
-        if self.use_ada_layer_norm_single:
-            norm_hidden_states = self.norm2_ln(hidden_states)
-            norm_hidden_states = norm_hidden_states * (1 + scale_mlp) + shift_mlp
-
-        ff_output = self.ff(norm_hidden_states, scale=lora_scale)
-
-        if self.use_ada_layer_norm_zero:
-            ff_output = gate_mlp.unsqueeze(1) * ff_output
-        elif self.use_ada_layer_norm_single:
-            ff_output = gate_mlp * ff_output
-
-        hidden_states = ff_output + hidden_states
-        if hidden_states.ndim == 4:
-            hidden_states = hidden_states.squeeze(1)
-
-        return hidden_states
-
-
-class Latte(ModelMixin, ConfigMixin):
-    _supports_gradient_checkpointing = True
-
-    """
-    A 2D Transformer model for image-like data.
-
-    Parameters:
-        num_attention_heads (`int`, *optional*, defaults to 16): The number of heads to use for multi-head attention.
-        attention_head_dim (`int`, *optional*, defaults to 88): The number of channels in each head.
-        in_channels (`int`, *optional*):
-            The number of channels in the input and output (specify if the input is **continuous**).
-        num_layers (`int`, *optional*, defaults to 1): The number of layers of Transformer blocks to use.
-        dropout (`float`, *optional*, defaults to 0.0): The dropout probability to use.
-        cross_attention_dim (`int`, *optional*): The number of `encoder_hidden_states` dimensions to use.
-        sample_size (`int`, *optional*): The width of the latent images (specify if the input is **discrete**).
-            This is fixed during training since it is used to learn a number of position embeddings.
-        num_vector_embeds (`int`, *optional*):
-            The number of classes of the vector embeddings of the latent pixels (specify if the input is **discrete**).
-            Includes the class for the masked latent pixel.
-        activation_fn (`str`, *optional*, defaults to `"geglu"`): Activation function to use in feed-forward.
-        num_embeds_ada_norm ( `int`, *optional*):
-            The number of diffusion steps used during training. Pass if at least one of the norm_layers is
-            `AdaLayerNorm`. This is fixed during training since it is used to learn a number of embeddings that are
-            added to the hidden states.
-
-            During inference, you can denoise for up to but not more steps than `num_embeds_ada_norm`.
-        attention_bias (`bool`, *optional*):
-            Configure if the `TransformerBlocks` attention should contain a bias parameter.
-    """
-
-    @register_to_config
-    def __init__(
-        self,
-        num_attention_heads: int = 16,
-        patch_size_t: int = 1,
-        attention_head_dim: int = 88,
-        in_channels: Optional[int] = None,
-        out_channels: Optional[int] = None,
-        num_layers: int = 1,
-        dropout: float = 0.0,
-        norm_num_groups: int = 32,
-        cross_attention_dim: Optional[int] = None,
-        attention_bias: bool = False,
-        sample_size: Optional[int] = None,
-        num_vector_embeds: Optional[int] = None,
-        patch_size: Optional[int] = None,
-        activation_fn: str = "geglu",
-        num_embeds_ada_norm: Optional[int] = None,
-        use_linear_projection: bool = False,
-        only_cross_attention: bool = False,
-        double_self_attention: bool = False,
-        upcast_attention: bool = False,
-        norm_type: str = "layer_norm",
-        norm_elementwise_affine: bool = True,
-        norm_eps: float = 1e-5,
-        attention_type: str = "default",
-        caption_channels: int = None,
-        video_length: int = 16,
-        enable_flash_attention: bool = False,
-        dtype=ms.float32,
-        use_recompute=False,
-    ):
-        super().__init__()
-        # self.use_linear_projection = use_linear_projection
-        self.num_attention_heads = num_attention_heads
-        self.attention_head_dim = attention_head_dim
-        inner_dim = num_attention_heads * attention_head_dim
-        self.video_length = video_length
-        self.norm_type = norm_type
-        self.use_recompute = use_recompute
-
-        conv_cls = nn.Conv2d  # if USE_PEFT_BACKEND else LoRACompatibleConv
-        linear_cls = nn.Dense  # if USE_PEFT_BACKEND else LoRACompatibleLinear
-
-        # 1. Transformer2DModel can process both standard continuous images of shape \
-        # `(batch_size, num_channels, width, height)` as well as quantized image embeddings of shape `(batch_size, num_image_vectors)`
-        # Define whether input is continuous or discrete depending on configuration
-        self.is_input_continuous = (in_channels is not None) and (patch_size is None)
-        self.is_input_vectorized = num_vector_embeds is not None
-        self.is_input_patches = in_channels is not None and patch_size is not None
-
-        if norm_type == "layer_norm" and num_embeds_ada_norm is not None:
-            deprecation_message = (
-                f"The configuration file of this model: {self.__class__} is outdated. `norm_type` is either not set or"
-                " incorrectly set to `'layer_norm'`.Make sure to set `norm_type` to `'ada_norm'` in the config."
-                " Please make sure to update the config accordingly as leaving `norm_type` might led to incorrect"
-                " results in future versions. If you have downloaded this checkpoint from the Hugging Face Hub, it"
-                " would be very nice if you could open a Pull request for the `transformer/config.json` file"
-            )
-            logger.warning("norm_type!=num_embeds_ada_norm", "1.0.0", deprecation_message, standard_warn=False)
-            norm_type = "ada_norm"
-
-        if self.is_input_continuous and self.is_input_vectorized:
-            raise ValueError(
-                f"Cannot define both `in_channels`: {in_channels} and `num_vector_embeds`: {num_vector_embeds}. Make"
-                " sure that either `in_channels` or `num_vector_embeds` is None."
-            )
-        elif self.is_input_vectorized and self.is_input_patches:
-            raise ValueError(
-                f"Cannot define both `num_vector_embeds`: {num_vector_embeds} and `patch_size`: {patch_size}. Make"
-                " sure that either `num_vector_embeds` or `num_patches` is None."
-            )
-        elif not self.is_input_continuous and not self.is_input_vectorized and not self.is_input_patches:
-            raise ValueError(
-                f"Has to define `in_channels`: {in_channels}, `num_vector_embeds`: {num_vector_embeds}, or patch_size:"
-                f" {patch_size}. Make sure that `in_channels`, `num_vector_embeds` or `num_patches` is not None."
-            )
-
-        # 2. Define input layers
-        if self.is_input_continuous:
-            self.in_channels = in_channels
-
-            self.norm = nn.GroupNorm(num_groups=norm_num_groups, num_channels=in_channels, eps=1e-6, affine=True)
-            if use_linear_projection:
-                self.proj_in = linear_cls(in_channels, inner_dim)
-            else:
-                self.proj_in = conv_cls(in_channels, inner_dim, kernel_size=1, stride=1, padding=0)
-        elif self.is_input_vectorized:
-            assert sample_size is not None, "Transformer2DModel over discrete input must provide sample_size"
-            assert num_vector_embeds is not None, "Transformer2DModel over discrete input must provide num_embed"
-
-            self.height = sample_size[0]
-            self.width = sample_size[1]
-            self.num_vector_embeds = num_vector_embeds
-            self.num_latent_pixels = self.height * self.width
-
-            self.latent_image_embedding = ImagePositionalEmbeddings(
-                num_embed=num_vector_embeds, embed_dim=inner_dim, height=self.height, width=self.width
-            )
-        elif self.is_input_patches:
-            assert sample_size is not None, "Transformer2DModel over patched input must provide sample_size"
-
-            self.height = sample_size[0]
-            self.width = sample_size[1]
-
-            self.patch_size = patch_size
-            interpolation_scale = self.config.sample_size[0] // 64  # => 64 (= 512 pixart) has interpolation scale 1
-            interpolation_scale = max(interpolation_scale, 1)
-            self.pos_embed = PatchEmbed(
-                height=sample_size[0],
-                width=sample_size[1],
-                patch_size=patch_size,
-                in_channels=in_channels,
-                embed_dim=inner_dim,
-                interpolation_scale=interpolation_scale,
-            )
-
-        # 3. Define transformers blocks
-        self.transformer_blocks = nn.CellList(
-            [
-                BasicTransformerBlock_(
-                    inner_dim,
-                    num_attention_heads,
-                    attention_head_dim,
-                    dropout=dropout,
-                    cross_attention_dim=None,  # unconditon do not need cross attn
-                    activation_fn=activation_fn,
-                    num_embeds_ada_norm=num_embeds_ada_norm,
-                    attention_bias=attention_bias,
-                    only_cross_attention=only_cross_attention,
-                    double_self_attention=False,
-                    upcast_attention=upcast_attention,
-                    norm_type=norm_type,
-                    norm_elementwise_affine=norm_elementwise_affine,
-                    norm_eps=norm_eps,
-                    attention_type=attention_type,
-                    enable_flash_attention=enable_flash_attention,
-                )
-                for d in range(num_layers)
-            ]
-        )
-
-        # Define temporal transformers blocks
-        self.temporal_transformer_blocks = nn.CellList(
-            [
-                BasicTransformerBlock_(  # one attention
-                    inner_dim,
-                    num_attention_heads,  # num_attention_heads
-                    attention_head_dim,  # attention_head_dim 72
-                    dropout=dropout,
-                    cross_attention_dim=None,
-                    activation_fn=activation_fn,
-                    num_embeds_ada_norm=num_embeds_ada_norm,
-                    attention_bias=attention_bias,
-                    only_cross_attention=only_cross_attention,
-                    double_self_attention=False,
-                    upcast_attention=upcast_attention,
-                    norm_type=norm_type,
-                    norm_elementwise_affine=norm_elementwise_affine,
-                    norm_eps=norm_eps,
-                    attention_type=attention_type,
-                    enable_flash_attention=enable_flash_attention,
-                )
-                for d in range(num_layers)
-            ]
-        )
-
-        # 4. Define output layers
-        self.out_channels = in_channels if out_channels is None else out_channels
-        if self.is_input_continuous:
-            # TODO: should use out_channels for continuous projections
-            if use_linear_projection:
-                self.proj_out = linear_cls(inner_dim, in_channels)
-            else:
-                self.proj_out = conv_cls(inner_dim, in_channels, kernel_size=1, stride=1, padding=0)
-        elif self.is_input_vectorized:
-            self.norm_out = LayerNorm(inner_dim)
-            self.out = nn.Dense(inner_dim, self.num_vector_embeds - 1)
-        elif self.is_input_patches and norm_type != "ada_norm_single":
-            self.norm_out = LayerNorm(inner_dim, elementwise_affine=False, eps=1e-6)
-            self.proj_out_1 = nn.Dense(inner_dim, 2 * inner_dim)
-            self.proj_out_2 = nn.Dense(inner_dim, patch_size * patch_size * self.out_channels)
-        elif self.is_input_patches and norm_type == "ada_norm_single":
-            self.norm_out = LayerNorm(inner_dim, elementwise_affine=False, eps=1e-6)
-            self.scale_shift_table = ms.Parameter(ops.randn(2, inner_dim) / inner_dim**0.5)
-            self.proj_out = nn.Dense(inner_dim, patch_size * patch_size * self.out_channels)
-
-        # 5. PixArt-Alpha blocks.
-        self.adaln_single = None
-        self.use_additional_conditions = False
-        if norm_type == "ada_norm_single":
-            # self.use_additional_conditions = self.config.sample_size[0] == 128  # False, 128 -> 1024
-            # TODO(Sayak, PVP) clean this, for now we use sample size to determine whether to use
-            # additional conditions until we find better name
-            self.adaln_single = AdaLayerNormSingle(inner_dim, use_additional_conditions=self.use_additional_conditions)
-
-        self.caption_projection = None
-        if caption_channels is not None:
-            self.caption_projection = CaptionProjection(in_features=caption_channels, hidden_size=inner_dim)
-
-        self.gradient_checkpointing = False
-
-        interpolation_scale = self.config.video_length // 5  # => 5 (= 5 our causalvideovae) has interpolation scale 1
-        interpolation_scale = max(interpolation_scale, 1)
-        temp_pos_embed = get_1d_sincos_pos_embed(
-            inner_dim, video_length, interpolation_scale=interpolation_scale
-        )  # 1152 hidden size
-        self.temp_pos_embed = Parameter(ms.Tensor(temp_pos_embed).float().unsqueeze(0), requires_grad=False)
-
-        if self.use_recompute:
-            for block in self.transformer_blocks:
-                self.recompute(block)
-
-            for block in self.temporal_transformer_blocks:
-                self.recompute(block)
-
-    def construct(
-        self,
-        hidden_states: ms.Tensor,
-        timestep: Optional[ms.Tensor] = None,
-        encoder_hidden_states: Optional[ms.Tensor] = None,
-        added_cond_kwargs: Dict[str, ms.Tensor] = None,
-        class_labels: Optional[ms.Tensor] = None,
-        cross_attention_kwargs: Dict[str, Any] = None,
-        attention_mask: Optional[ms.Tensor] = None,
-        encoder_attention_mask: Optional[ms.Tensor] = None,
-        use_image_num: int = 0,
-        enable_temporal_attentions: bool = True,
-    ):
-        """
-        The [`Transformer2DModel`] forward method.
-
-        Args:
-            hidden_states (`ms.Tensor` of shape `(batch size, num latent pixels)` if discrete, `ms.Tensor` of shape \
-                `(batch size, frame, channel, height, width)` if continuous): Input `hidden_states`.
-            encoder_hidden_states ( `ms.Tensor` of shape `(batch size, sequence len, embed dims)`, *optional*):
-                Conditional embeddings for cross attention layer. If not given, cross-attention defaults to
-                self-attention.
-            timestep ( `ms.Tensor`, *optional*):
-                Used to indicate denoising step. Optional timestep to be applied as an embedding in `AdaLayerNorm`.
-            class_labels ( `ms.Tensor` of shape `(batch size, num classes)`, *optional*):
-                Used to indicate class labels conditioning. Optional class labels to be applied as an embedding in
-                `AdaLayerZeroNorm`.
-            cross_attention_kwargs ( `Dict[str, Any]`, *optional*):
-                A kwargs dictionary that if specified is passed along to the `AttentionProcessor` as defined under
-                `self.processor` in
-                [diffusers.models.attention_processor](https://github.com/huggingface/diffusers/blob/main/src/diffusers/models/attention_processor.py).
-            attention_mask ( `ms.Tensor`, *optional*):
-                An attention mask of shape `(batch, key_tokens)` is applied to `encoder_hidden_states`. If `1` the mask
-                is kept, otherwise if `0` it is discarded. Mask will be converted into a bias, which adds large
-                negative values to the attention scores corresponding to "discard" tokens.
-            encoder_attention_mask ( `ms.Tensor`, *optional*):
-                Cross-attention mask applied to `encoder_hidden_states`. Two formats supported:
-
-                    * Mask `(batch, sequence_length)` True = keep, False = discard.
-                    * Bias `(batch, 1, sequence_length)` 0 = keep, -10000 = discard.
-
-                If `ndim == 2`: will be interpreted as a mask, then converted into a bias consistent with the format
-                above. This bias will be added to the cross-attention scores.
-        """
-        input_batch_size, c, frame, h, w = hidden_states.shape
-        frame = frame - use_image_num
-        # b c f h w -> (b f) c h w
-        hidden_states = hidden_states.permute(0, 2, 1, 3, 4).view(-1, c, h, w)
-
-        # ensure attention_mask is a bias, and give it a singleton query_tokens dimension.
-        #   we may have done this conversion already, e.g. if we came here via UNet2DConditionModel#forward.
-        #   we can tell by counting dims; if ndim == 2: it's a mask rather than a bias.
-        # expects mask of shape:
-        #   [batch, key_tokens]
-        # adds singleton query_tokens dimension:
-        #   [batch,                    1, key_tokens]
-        # this helps to broadcast it as a bias over attention scores, which will be in one of the following shapes:
-        #   [batch,  heads, query_tokens, key_tokens] (e.g. torch sdp attn)
-        #   [batch * heads, query_tokens, key_tokens] (e.g. xformers or classic attn)
-        if attention_mask is not None and attention_mask.ndim == 2:
-            # assume that mask is expressed as:
-            #   (1 = keep,      0 = discard)
-            # convert mask into a bias that can be added to attention scores:
-            #       (keep = +0,     discard = -ms.numpy.inf)
-            attention_mask = attention_mask.unsqueeze(1)  # (b, 1, key_len)
-            attention_mask = ops.zeros(attention_mask.shape).masked_fill(~attention_mask, -ms.numpy.inf)
-
-        # # Retrieve lora scale.
-        # lora_scale = cross_attention_kwargs.get("scale", 1.0) if cross_attention_kwargs is not None else 1.0
-
-        # 1. Input
-        assert self.is_input_patches, "Currently only support input patches!"
-        # if self.is_input_patches:  # here
-        height, width = hidden_states.shape[-2] // self.patch_size, hidden_states.shape[-1] // self.patch_size
-        num_patches = height * width
-
-        hidden_states = self.pos_embed(hidden_states.to(self.dtype))  # alrady add positional embeddings
-
-        if self.adaln_single is not None:
-            if self.use_additional_conditions and added_cond_kwargs is None:
-                raise ValueError(
-                    "`added_cond_kwargs` cannot be None when using additional conditions for `adaln_single`."
-                )
-            # batch_size = hidden_states.shape[0]
-            batch_size = input_batch_size
-            timestep, embedded_timestep = self.adaln_single(
-                timestep, added_cond_kwargs, batch_size=batch_size, hidden_dtype=hidden_states.dtype
-            )
-        else:
-            embedded_timestep = None
-
-        # prepare timesteps for spatial and temporal block
-        # b d -> (b f) d
-        timestep_spatial = timestep.repeat_interleave(frame + use_image_num, dim=0)
-        # b d -> (b p) d
-        timestep_temp = timestep.repeat_interleave(num_patches, dim=0)
-
-        for i, (spatial_block, temp_block) in enumerate(zip(self.transformer_blocks, self.temporal_transformer_blocks)):
-            hidden_states = spatial_block(
-                hidden_states,
-                attention_mask,
-                None,  # encoder_hidden_states_spatial
-                None,  # encoder_attention_mask
-                timestep_spatial,
-                cross_attention_kwargs,
-                class_labels,
-            )
-
-            if enable_temporal_attentions:
-                # (b f) t d -> (b t) f d
-                hidden_states = hidden_states.view(
-                    input_batch_size, frame + use_image_num, hidden_states.shape[1], hidden_states.shape[2]
-                )
-                hidden_states = hidden_states.permute(0, 2, 1, 3).view(
-                    -1, frame + use_image_num, hidden_states.shape[-1]
-                )
-
-                if use_image_num != 0 and self.training:
-                    hidden_states_video = hidden_states[:, :frame, ...]
-                    hidden_states_image = hidden_states[:, frame:, ...]
-
-                    hidden_states_video = temp_block(
-                        hidden_states_video,
-                        None,  # attention_mask
-                        None,  # encoder_hidden_states
-                        None,  # encoder_attention_mask
-                        timestep_temp,
-                        cross_attention_kwargs,
-                        class_labels,
-                    )
-
-                    hidden_states = ops.cat([hidden_states_video, hidden_states_image], axis=1)
-                    # (b t) f d -> (b f) t d
-                    hidden_states = hidden_states.view(
-                        input_batch_size, -1, hidden_states.shape[1], hidden_states.shape[2]
-                    )
-                    hidden_states = hidden_states.permute(0, 2, 1, 3).view(
-                        input_batch_size * (frame + use_image_num), -1, hidden_states.shape[-1]
-                    )
-
-                else:
-                    if i == 0:
-                        hidden_states = hidden_states + self.temp_pos_embed
-
-                    hidden_states = temp_block(
-                        hidden_states,
-                        None,  # attention_mask
-                        None,  # encoder_hidden_states
-                        None,  # encoder_attention_mask
-                        timestep_temp,
-                        cross_attention_kwargs,
-                        class_labels,
-                    )
-
-                    # (b t) f d -> (b f) t d
-                    hidden_states = hidden_states.view(
-                        input_batch_size, -1, hidden_states.shape[1], hidden_states.shape[2]
-                    )
-                    hidden_states = hidden_states.permute(0, 2, 1, 3).view(
-                        input_batch_size * (frame + use_image_num), -1, hidden_states.shape[-1]
-                    )
-
-        # if self.is_input_patches:
-        if self.norm_type != "ada_norm_single":
-            conditioning = self.transformer_blocks[0].norm1.emb(
-                timestep, class_labels, hidden_dtype=hidden_states.dtype
-            )
-            shift, scale = self.proj_out_1(ops.silu(conditioning)).chunk(2, axis=1)
-            hidden_states = self.norm_out(hidden_states) * (1 + scale[:, None]) + shift[:, None]
-            hidden_states = self.proj_out_2(hidden_states)
-        elif self.norm_type == "ada_norm_single":
-            # b d -> (b f) d
-            assert embedded_timestep is not None, "embedded_timestep should be not None"
-            embedded_timestep = embedded_timestep.repeat_interleave(frame + use_image_num, dim=0)
-            shift, scale = (self.scale_shift_table[None] + embedded_timestep[:, None]).chunk(2, axis=1)
-            hidden_states = self.norm_out(hidden_states)
-            # Modulation
-            hidden_states = hidden_states * (1 + scale) + shift
-            hidden_states = self.proj_out(hidden_states)
-
-        # unpatchify
-        if self.adaln_single is None:
-            height = width = int(hidden_states.shape[1] ** 0.5)
-        hidden_states = hidden_states.reshape(
-            shape=(-1, height, width, self.patch_size, self.patch_size, self.out_channels)
-        )
-        # nhwpqc->nchpwq
-        hidden_states = hidden_states.permute(0, 5, 1, 3, 2, 4)
-        output = hidden_states.reshape(shape=(-1, self.out_channels, height * self.patch_size, width * self.patch_size))
-        # (b f) c h w -> b c f h w
-        output = output.view(input_batch_size, -1, output.shape[-3], output.shape[-2], output.shape[-1])
-        output = output.permute(0, 2, 1, 3, 4)
-
-        return output
-
-    @classmethod
-    def from_pretrained_2d(cls, pretrained_model_path, subfolder=None, **kwargs):
-        if subfolder is not None:
-            pretrained_model_path = os.path.join(pretrained_model_path, subfolder)
-
-        config_file = os.path.join(pretrained_model_path, "config.json")
-        if not os.path.isfile(config_file):
-            raise RuntimeError(f"{config_file} does not exist")
-        with open(config_file, "r") as f:
-            config = json.load(f)
-
-        model = cls.from_config(config, **kwargs)
-        return model
-
-    def construct_with_cfg(self, x, timestep, class_labels=None, cfg_scale=7.0, attention_mask=None):
-        """
-        Forward pass of Latte, but also batches the unconditional forward pass for classifier-free guidance.
-        """
-        # https://github.com/openai/glide-text2im/blob/main/notebooks/text2im.ipynb
-        # x shape : b f c h w
-        half = x[: len(x) // 2]
-        combined = ops.cat([half, half], axis=0)
-        model_out = self.construct(combined, timestep, class_labels=class_labels, attention_mask=attention_mask)
-        # For exact reproducibility reasons, we apply classifier-free guidance on only
-        # three channels by default. The standard approach to cfg applies it to all channels.
-        # This can be done by uncommenting the following line and commenting-out the line following that.
-        eps, rest = model_out[:, :, : self.in_channels], model_out[:, :, self.in_channels :]
-        cond_eps, uncond_eps = ops.split(eps, len(eps) // 2, axis=0)
-        half_eps = uncond_eps + cfg_scale * (cond_eps - uncond_eps)
-        eps = ops.cat([half_eps, half_eps], axis=0)
-        return ops.cat([eps, rest], axis=2)
-
-    def load_params_from_ckpt(self, ckpt):
-        # load param from a ckpt file path or a parameter dictionary
-        if isinstance(ckpt, str):
-            assert os.path.exists(ckpt), f"{ckpt} does not exist!"
-            logger.info(f"Loading {ckpt} params into Latte_T2V_SD model...")
-            param_dict = ms.load_checkpoint(ckpt)
-        elif isinstance(ckpt, dict):
-            param_dict = ckpt
-        else:
-            raise ValueError("Expect to receive a ckpt path or parameter dictionary as input!")
-        _, ckpt_not_load = ms.load_param_into_net(
-            self,
-            param_dict,
-        )
-        if len(ckpt_not_load):
-            print(f"{ckpt_not_load} not load")
-
-    def recompute(self, b):
-        if not b._has_config_recompute:
-            b.recompute()
-        if isinstance(b, nn.CellList):
-            self.recompute(b[-1])
-        else:
-            b.add_flags(output_no_recompute=True)
-
-
-class LatteT2VBlock(nn.Cell):
-    def __init__(self, block_id, temp_pos_embed, spatial_block, temp_block):
-        super().__init__()
-        self.spatial_block = spatial_block
-        self.temp_block = temp_block
-        self.is_first_block = block_id == 0
-        self.temp_pos_embed = temp_pos_embed
-
-    def construct(
-        self,
-        hidden_states: ms.Tensor,
-        class_labels: Optional[ms.Tensor] = None,
-        cross_attention_kwargs: Dict[str, Any] = None,
-        attention_mask: Optional[ms.Tensor] = None,
-        encoder_hidden_states_spatial: Optional[ms.Tensor] = None,
-        timestep_spatial: Optional[ms.Tensor] = None,
-        timestep_temp: Optional[ms.Tensor] = None,
-        encoder_attention_mask: Optional[ms.Tensor] = None,
-        use_image_num: int = 0,
-        input_batch_size: int = 0,
-        frame: int = 0,
-        enable_temporal_attentions: bool = True,
-    ):
-        hidden_states = self.spatial_block(
-            hidden_states,
-            attention_mask,
-            encoder_hidden_states_spatial,
-            encoder_attention_mask,
-            timestep_spatial,
-            cross_attention_kwargs,
-            class_labels,
-        )
-
-        if enable_temporal_attentions:
-            # b c f h w, f = 16 + 4
-            # (b f) t d -> (b t) f d
-            hidden_states = hidden_states.view(input_batch_size, frame + use_image_num, -1, hidden_states.shape[-1])
-            hidden_states = hidden_states.permute(0, 2, 1, 3).view(-1, frame + use_image_num, hidden_states.shape[-1])
-
-            if use_image_num != 0 and self.training:
-                hidden_states_video = hidden_states[:, :frame, ...]
-                hidden_states_image = hidden_states[:, frame:, ...]
-                if self.is_first_block:
-                    hidden_states_video = hidden_states_video + self.temp_pos_embed
-
-                hidden_states_video = self.temp_block(
-                    hidden_states_video,
-                    None,  # attention_mask
-                    None,  # encoder_hidden_states
-                    None,  # encoder_attention_mask
-                    timestep_temp,
-                    cross_attention_kwargs,
-                    class_labels,
-                )
-
-                hidden_states = ops.cat([hidden_states_video, hidden_states_image], axis=1)
-                # (b t) f d -> (b f) t d
-                hidden_states = hidden_states.view(input_batch_size, -1, frame + use_image_num, hidden_states.shape[-1])
-                hidden_states = hidden_states.permute(0, 2, 1, 3).view(
-                    input_batch_size * (frame + use_image_num), -1, hidden_states.shape[-1]
-                )
-
-            else:
-                if self.is_first_block:
-                    hidden_states = hidden_states + self.temp_pos_embed
-
-                hidden_states = self.temp_block(
-                    hidden_states,
-                    None,  # attention_mask
-                    None,  # encoder_hidden_states
-                    None,  # encoder_attention_mask
-                    timestep_temp,
-                    cross_attention_kwargs,
-                    class_labels,
-                )
-                # (b t) f d -> (b f) t d
-                hidden_states = hidden_states.view(input_batch_size, -1, frame + use_image_num, hidden_states.shape[-1])
-                hidden_states = hidden_states.permute(0, 2, 1, 3).view(
-                    input_batch_size * (frame + use_image_num), -1, hidden_states.shape[-1]
-                )
-        return hidden_states
 
 
 class LatteT2V(ModelMixin, ConfigMixin):
@@ -1165,26 +87,37 @@ class LatteT2V(ModelMixin, ConfigMixin):
         video_length: int = 16,
         enable_flash_attention: bool = False,
         use_recompute=False,
+        use_rope: bool = False,
+        model_max_length: int = 300,
+        rope_scaling_type: str = "linear",
+        compress_kv_factor: int = 1,
+        interpolation_scale_1d: float = None,
     ):
         super().__init__()
-        # self.use_linear_projection = use_linear_projection
+        self.use_linear_projection = use_linear_projection
         self.num_attention_heads = num_attention_heads
         self.attention_head_dim = attention_head_dim
         inner_dim = num_attention_heads * attention_head_dim
         self.video_length = video_length
         self.norm_type = norm_type
         self.use_recompute = use_recompute
+        self.use_rope = use_rope
+        self.model_max_length = model_max_length
+        self.compress_kv_factor = compress_kv_factor
+        self.num_layers = num_layers
+        self.config.hidden_size = model_max_length
+
+        assert not (self.compress_kv_factor != 1 and use_rope), "Can not both enable compressing kv and using rope"
 
         conv_cls = nn.Conv2d  # if USE_PEFT_BACKEND else LoRACompatibleConv
         linear_cls = nn.Dense  # if USE_PEFT_BACKEND else LoRACompatibleLinear
 
-        # 1. Transformer2DModel can process both standard continuous images of shape \
-        # `(batch_size, num_channels, width, height)` as well as quantized image embeddings of shape \
-        # `(batch_size, num_image_vectors)`
+        # 1. Transformer2DModel can process both standard continuous images of shape
         # Define whether input is continuous or discrete depending on configuration
         self.is_input_continuous = (in_channels is not None) and (patch_size is None)
         self.is_input_vectorized = num_vector_embeds is not None
-        self.is_input_patches = in_channels is not None and patch_size is not None
+        # self.is_input_patches = in_channels is not None and patch_size is not None
+        self.is_input_patches = True
 
         if norm_type == "layer_norm" and num_embeds_ada_norm is not None:
             deprecation_message = (
@@ -1197,62 +130,45 @@ class LatteT2V(ModelMixin, ConfigMixin):
             logger.warning("norm_type!=num_embeds_ada_norm", "1.0.0", deprecation_message, standard_warn=False)
             norm_type = "ada_norm"
 
-        if self.is_input_continuous and self.is_input_vectorized:
-            raise ValueError(
-                f"Cannot define both `in_channels`: {in_channels} and `num_vector_embeds`: {num_vector_embeds}. Make"
-                " sure that either `in_channels` or `num_vector_embeds` is None."
-            )
-        elif self.is_input_vectorized and self.is_input_patches:
-            raise ValueError(
-                f"Cannot define both `num_vector_embeds`: {num_vector_embeds} and `patch_size`: {patch_size}. Make"
-                " sure that either `num_vector_embeds` or `num_patches` is None."
-            )
-        elif not self.is_input_continuous and not self.is_input_vectorized and not self.is_input_patches:
-            raise ValueError(
-                f"Has to define `in_channels`: {in_channels}, `num_vector_embeds`: {num_vector_embeds}, or patch_size:"
-                f" {patch_size}. Make sure that `in_channels`, `num_vector_embeds` or `num_patches` is not None."
-            )
-
         # 2. Define input layers
-        if self.is_input_continuous:
-            self.in_channels = in_channels
+        assert sample_size is not None, "Transformer2DModel over patched input must provide sample_size"
 
-            self.norm = nn.GroupNorm(num_groups=norm_num_groups, num_channels=in_channels, eps=1e-6, affine=True)
-            if use_linear_projection:
-                self.proj_in = linear_cls(in_channels, inner_dim)
+        self.height = sample_size[0]
+        self.width = sample_size[1]
+
+        self.patch_size = patch_size
+        interpolation_scale_2d = self.config.sample_size[0] // 64  # => 64 (= 512 pixart) has interpolation scale 1
+        interpolation_scale_2d = max(interpolation_scale_2d, 1)
+        self.pos_embed = PatchEmbed(
+            height=sample_size[0],
+            width=sample_size[1],
+            patch_size=patch_size,
+            in_channels=in_channels,
+            embed_dim=inner_dim,
+            interpolation_scale=interpolation_scale_2d,
+        )
+
+        # define temporal positional embedding
+        if interpolation_scale_1d is None:
+            if self.config.video_length % 2 == 1:
+                interpolation_scale_1d = (
+                    self.config.video_length - 1
+                ) // 16  # => 16 (= 16 Latte) has interpolation scale 1
             else:
-                self.proj_in = conv_cls(in_channels, inner_dim, kernel_size=1, stride=1, padding=0)
-        elif self.is_input_vectorized:
-            assert sample_size is not None, "Transformer2DModel over discrete input must provide sample_size"
-            assert num_vector_embeds is not None, "Transformer2DModel over discrete input must provide num_embed"
+                interpolation_scale_1d = self.config.video_length // 16  # => 16 (= 16 Latte) has interpolation scale 1
+        # interpolation_scale_1d = self.config.video_length // 5  #
+        interpolation_scale_1d = max(interpolation_scale_1d, 1)
+        temp_pos_embed = get_1d_sincos_pos_embed(
+            inner_dim, video_length, interpolation_scale=interpolation_scale_1d
+        )  # 1152 hidden size
+        self.temp_pos_embed = ms.Parameter(ms.Tensor(temp_pos_embed).float().unsqueeze(0), requires_grad=False)
 
-            self.height = sample_size[0] if isinstance(sample_size, (tuple, list)) else sample_size
-            self.width = sample_size[1] if isinstance(sample_size, (tuple, list)) else sample_size
-            self.num_vector_embeds = num_vector_embeds
-            self.num_latent_pixels = self.height * self.width
-
-            self.latent_image_embedding = ImagePositionalEmbeddings(
-                num_embed=num_vector_embeds, embed_dim=inner_dim, height=self.height, width=self.width
-            )
-        elif self.is_input_patches:
-            assert sample_size is not None, "Transformer2DModel over patched input must provide sample_size"
-
-            self.height = sample_size[0] if isinstance(sample_size, (tuple, list)) else sample_size
-            self.width = sample_size[1] if isinstance(sample_size, (tuple, list)) else sample_size
-
-            self.patch_size = patch_size
-            if isinstance(self.config.sample_size, (tuple, list)):
-                interpolation_scale = self.config.sample_size[0] // 64  # => 64 (= 512 pixart) has interpolation scale 1
-            else:
-                interpolation_scale = self.config.sample_size // 64
-            interpolation_scale = max(interpolation_scale, 1)
-            self.pos_embed = PatchEmbed(
-                height=sample_size[0] if isinstance(sample_size, (tuple, list)) else sample_size,
-                width=sample_size[1] if isinstance(sample_size, (tuple, list)) else sample_size,
-                patch_size=patch_size,
-                in_channels=in_channels,
-                embed_dim=inner_dim,
-                interpolation_scale=interpolation_scale,
+        rope_scaling = None
+        if self.use_rope:
+            self.position_getter_2d = PositionGetter2D()
+            self.position_getter_1d = PositionGetter1D()
+            rope_scaling = dict(
+                type=rope_scaling_type, factor_2d=interpolation_scale_2d, factor_1d=interpolation_scale_1d
             )
 
         # 3. Define transformers blocks, spatial attention
@@ -1274,6 +190,11 @@ class LatteT2V(ModelMixin, ConfigMixin):
                 norm_eps=norm_eps,
                 attention_type=attention_type,
                 enable_flash_attention=enable_flash_attention,
+                use_rope=use_rope,
+                rope_scaling=rope_scaling,
+                compress_kv_factor=(compress_kv_factor, compress_kv_factor)
+                if d >= num_layers // 2 and compress_kv_factor != 1
+                else None,  # follow pixart-sigma, apply in second-half layers
             )
             for d in range(num_layers)
         ]
@@ -1297,6 +218,11 @@ class LatteT2V(ModelMixin, ConfigMixin):
                 norm_eps=norm_eps,
                 attention_type=attention_type,
                 enable_flash_attention=enable_flash_attention,
+                use_rope=use_rope,
+                rope_scaling=rope_scaling,
+                compress_kv_factor=(compress_kv_factor,)
+                if d >= num_layers // 2 and compress_kv_factor != 1
+                else None,  # follow pixart-sigma, apply in second-half layers
             )
             for d in range(num_layers)
         ]
@@ -1336,17 +262,6 @@ class LatteT2V(ModelMixin, ConfigMixin):
 
         self.gradient_checkpointing = False
 
-        # define temporal positional embedding
-        # temp_pos_embed = self.get_1d_sincos_temp_embed(inner_dim, video_length)  # 1152 hidden size
-
-        interpolation_scale = self.config.video_length // 5  # => 5 (= 5 our causalvideovae) has interpolation scale 1
-        interpolation_scale = max(interpolation_scale, 1)
-        temp_pos_embed = get_1d_sincos_pos_embed(
-            inner_dim, video_length, interpolation_scale=interpolation_scale
-        )  # 1152 hidden size
-
-        self.temp_pos_embed = ms.Tensor(temp_pos_embed).float().unsqueeze(0)
-
         self.blocks = nn.CellList(
             [
                 LatteT2VBlock(d, self.temp_pos_embed, self.transformer_blocks[d], self.temporal_transformer_blocks[d])
@@ -1357,6 +272,33 @@ class LatteT2V(ModelMixin, ConfigMixin):
         if self.use_recompute:
             for block in self.blocks:
                 self.recompute(block)
+
+    def make_position(self, b, t, use_image_num, h, w):
+        pos_hw = self.position_getter_2d(b * (t + use_image_num), h, w)  # fake_b = b*(t+use_image_num)
+        pos_t = self.position_getter_1d(b * h * w, t)  # fake_b = b*h*w
+        return pos_hw, pos_t
+
+    def make_attn_mask(self, attention_mask, frame, dtype):
+        # attention_mask = rearrange(attention_mask, 'b t h w -> (b t) 1 (h w)')
+        b, t, h, w = attention_mask.shape
+        attention_mask = attention_mask.reshape(b * t, h * w).unsqueeze(1)
+        # assume that mask is expressed as:
+        #   (1 = keep,      0 = discard)
+        # convert mask into a bias that can be added to attention scores:
+        #   (keep = +0,     discard = -10000.0)
+        attention_mask = (1 - attention_mask.to(dtype)) * -10000.0
+        attention_mask = attention_mask.to(self.dtype)
+        return attention_mask
+
+    def vae_to_diff_mask(self, attention_mask, use_image_num):
+        dtype = attention_mask.dtype
+        # b, t+use_image_num, h, w, assume t as channel
+        # this version do not use 3d patch embedding
+        attention_mask = ops.max_pool2d(
+            attention_mask, kernel_size=(self.patch_size, self.patch_size), stride=(self.patch_size, self.patch_size)
+        )
+        attention_mask = attention_mask.bool().to(dtype)
+        return attention_mask
 
     def construct(
         self,
@@ -1401,13 +343,10 @@ class LatteT2V(ModelMixin, ConfigMixin):
 
                 If `ndim == 2`: will be interpreted as a mask, then converted into a bias consistent with the format
                 above. This bias will be added to the cross-attention scores.
-
+        Returns:
+           a `tuple` where the first element is the sample tensor.
         """
         input_batch_size, c, frame, h, w = hidden_states.shape
-        # print(hidden_states.shape, input_batch_size, c, frame, h, w, use_image_num)
-        # print(timestep)
-        # print(encoder_hidden_states.shape)
-        # print(encoder_attention_mask.shape)
         frame = frame - use_image_num  # 20-4=16
         # b c f h w -> (b f) c h w
         hidden_states = hidden_states.permute(0, 2, 1, 3, 4).reshape(
@@ -1423,11 +362,18 @@ class LatteT2V(ModelMixin, ConfigMixin):
         # this helps to broadcast it as a bias over attention scores, which will be in one of the following shapes:
         #   [batch,  heads, query_tokens, key_tokens] (e.g. torch sdp attn)
         #   [batch * heads, query_tokens, key_tokens] (e.g. xformers or classic attn)
-        if attention_mask is not None and attention_mask.ndim == 2:
-            # assume that mask is expressed as:
-            #   (1 = keep,      0 = discard)
-            attention_mask = attention_mask.unsqueeze(1)
-            attention_mask = attention_mask.to(self.dtype)
+        if attention_mask is None:
+            attention_mask = ops.ones((input_batch_size, frame + use_image_num, h, w), dtype=hidden_states.dtype)
+        attention_mask = self.vae_to_diff_mask(attention_mask, use_image_num)
+        dtype = attention_mask.dtype
+        attention_mask_compress = ops.max_pool2d(
+            attention_mask.float(), kernel_size=self.compress_kv_factor, stride=self.compress_kv_factor
+        )
+        attention_mask_compress = attention_mask_compress.to(dtype)
+
+        attention_mask = self.make_attn_mask(attention_mask, frame, hidden_states.dtype)
+        attention_mask_compress = self.make_attn_mask(attention_mask_compress, frame, hidden_states.dtype)
+
         # 1 + 4, 1 -> video condition, 4 -> image condition
         # convert encoder_attention_mask to a bias the same way we do for attention_mask
         if encoder_attention_mask is not None and encoder_attention_mask.ndim == 2:  # ndim == 2 means no image joint
@@ -1451,6 +397,7 @@ class LatteT2V(ModelMixin, ConfigMixin):
         assert self.is_input_patches, "Only support input patches now!"
         # if self.is_input_patches:  # here
         height, width = hidden_states.shape[-2] // self.patch_size, hidden_states.shape[-1] // self.patch_size
+        hw = (height, width)
         num_patches = height * width
 
         hidden_states = self.pos_embed(hidden_states.to(self.dtype))  # alrady add positional embeddings
@@ -1495,12 +442,16 @@ class LatteT2V(ModelMixin, ConfigMixin):
         # b d -> (b p) d
         timestep_temp = timestep.repeat_interleave(num_patches, dim=0)
 
-        for block in self.blocks:
+        pos_hw, pos_t = None, None
+        if self.use_rope:
+            pos_hw, pos_t = self.make_position(input_batch_size, frame, use_image_num, height, width)
+
+        for i, block in enumerate(self.blocks):
             hidden_states = block(
                 hidden_states,
                 class_labels,
                 cross_attention_kwargs,
-                attention_mask,
+                attention_mask_compress if i >= self.num_layers // 2 else attention_mask,
                 encoder_hidden_states_spatial,
                 timestep_spatial,
                 timestep_temp,
@@ -1509,6 +460,9 @@ class LatteT2V(ModelMixin, ConfigMixin):
                 input_batch_size,
                 frame,
                 enable_temporal_attentions,
+                pos_hw=pos_hw,
+                pos_t=pos_t,
+                hw=hw,
             )
 
         # if self.is_input_patches:
@@ -1627,31 +581,6 @@ class LatteT2V(ModelMixin, ConfigMixin):
             print("ckpt param not load: ", u, len(u))
 
 
-# depth = num_layers * 2
-def Latte_XL_122(**kwargs):
-    return Latte(
-        num_layers=28,
-        attention_head_dim=72,
-        num_attention_heads=16,
-        patch_size_t=1,
-        patch_size=2,
-        norm_type="ada_norm_single",
-        **kwargs,
-    )
-
-
-def LatteClass_XL_122(**kwargs):
-    return Latte(
-        num_layers=28,
-        attention_head_dim=72,
-        num_attention_heads=16,
-        patch_size_t=1,
-        patch_size=2,
-        norm_type="ada_norm_zero",
-        **kwargs,
-    )
-
-
 def LatteT2V_XL_122(**kwargs):
     return LatteT2V(
         num_layers=28,
@@ -1667,7 +596,5 @@ def LatteT2V_XL_122(**kwargs):
 
 
 Latte_models = {
-    "Latte-XL/122": Latte_XL_122,
-    "LatteClass-XL/122": LatteClass_XL_122,
     "LatteT2V-XL/122": LatteT2V_XL_122,
 }
